create zip file loader
