--- downloader/downloadGCS.go
@@ -1,44 +1,0 @@
-package downloader
-
-import (
-	"cloud.google.com/go/storage"
-	"errors"
-	"golang.org/x/net/context"
-	"log"
-
-	"github.com/m-lab/annotation-service/parser"
-)
-
-var geoData []parser.Node
-
-//Creates list of IP address Nodes
-func InitializeTable(ctx context.Context, GCSFolder, GCSFile string, IPVersion int) ([]parser.Node, error) {
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	storageReader, err := createReader(GCSFolder, GCSFile, ctx)
-	if err != nil {
-		log.Println("storage reader failed")
-		return geoData, errors.New("Storage Reader Failed")
-	}
-	geoData, err = parser.CreateList(storageReader, IPVersion)
-	if err != nil {
-		log.Println("geoData createList failed")
-		return geoData, errors.New("geoData createList Failed")
-	}
-	return geoData, nil
-}
-
-//creates generic reader
-func createReader(bucket string, bucketObj string, ctx context.Context) (*storage.Reader, error) {
-	client, err := storage.NewClient(ctx)
-	if err != nil {
-		log.Fatal(err)
-	}
-	obj := client.Bucket(bucket).Object(bucketObj)
-	reader, err := obj.NewReader(ctx)
-	if err != nil {
-		log.Fatal(err)
-	}
-	return reader, nil
-}
