--- downloader/testdata/GeoIPCountryWhoisSAMPLE.csv
@@ -1,4 +1,0 @@
-"1.0.1.0","1.0.3.255","16777472","16778239","CN","China"
-"1.0.4.0","1.0.7.255","16778240","16779263","AU","Australia"
-"1.0.8.0","1.0.15.255","16779264","16781311","CN","China"
-"1.0.16.0","1.0.31.255","16781312","16785407","JP","Japan"
