--- downloader/testdata/GeoLiteCityv6SAMPLE.csv
@@ -1,4 +1,0 @@
-2001:5::,2001:5:ffff:ffff:ffff:ffff:ffff:ffff,42540488558116655331872044393019998208,42540488637344817846136381986563948543,EU,,,47.0000,8.0000,,0,0
-2001:200::,2001:200:ffff:ffff:ffff:ffff:ffff:ffff,42540528726795050063891204319802818560,42540528806023212578155541913346768895,JP,,,36.0000,138.0000,,0,0
-2001:208::,2001:208:ffff:ffff:ffff:ffff:ffff:ffff,42540529360620350178005905068154421248,42540529439848512692270242661698371583,SG,,,1.3667,103.8000,,0,0
-2001:218::,2001:218:ffff:ffff:ffff:ffff:ffff:ffff,42540530628270950406235306564857626624,42540530707499112920499644158401576959,JP,,,36.0000,138.0000,,0,0
